requirements.txt
appdirs==1.4.0
Django==1.11.21
packaging==16.8
pyparsing==2.1.10
six==1.10.0
django-polymorphic==2.1.2
pillow==6.1.0
